// src/lib/api/generated/models/problemResponseStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Campus Ativo API
 * API para gerenciamento de problemas de infraestrutura
 * OpenAPI spec version: 1.0.0
 */

/**
 * Status do problema
 */
export type ProblemResponseStatus = typeof ProblemResponseStatus[keyof typeof ProblemResponseStatus];


export const ProblemResponseStatus = {
  TO_ANALYSIS: 'TO_ANALYSIS',
  IN_ANALYSIS: 'IN_ANALYSIS',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
  IN_PROGRESS: 'IN_PROGRESS',
  FINISHED: 'FINISHED',
} as const;